// types/models.ts
export type TypeSlideModel = {
  _id: string;
  name: string;
  description: string;
  slug: string;
  image: string;
  user_id: string;
  status: "draft" | "publish" | "archive";
  slideItem: string[];
  createdAt: Date;
};